# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MyWebDrive Public API
  version: 1.0.0
  description: |
    Public Core-first contract served through the same-origin Nginx listener.
    Core owns identity, administrator operations, files, upload intents, quota,
    sharing, publications, notifications, and dashboards. Storage accepts only
    short-lived grants issued by Core.

    Private callbacks under `/api/v1/internal/*` and operational endpoints
    `/metrics`, `/live`, `/ready`, and `/version` are intentionally excluded.
  license:
    name: MIT
    url: https://opensource.org/license/mit
...
servers:
  - url: http://127.0.0.1:8080
    description: Core-first local development listener
...
paths:
...
  /api/v1/admin/users:
    get:
      operationId: listAdminUsers
      tags: [Admin]
      summary: List Core identities with authoritative quota balances
      security:
        - accessToken: []
      parameters:
        - name: query
          in: query
          required: false
          schema: { type: string, maxLength: 200 }
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Page of Core users ordered newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUserPage' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/admin/users/{userId}:
    get:
      operationId: getAdminUser
      tags: [Admin]
      summary: Return one Core identity with its quota balance
      security:
        - accessToken: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Core user and quota balance.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUser' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/admin/users/{userId}/role:
    patch:
      operationId: updateAdminUserRole
      tags: [Admin]
      summary: Change a Core user's role
      security:
        - accessToken: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role: { type: string, enum: [user, admin] }
      responses:
        '200':
          description: Updated role.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminRoleUpdate' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/admin/notifications:
    get:
      operationId: listAdminNotifications
      tags: [Admin]
      summary: List persistent Core administrator notifications
      security:
        - accessToken: []
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema: { type: boolean, default: false }
        - name: service
          in: query
          required: false
          schema: { type: string, maxLength: 100 }
        - name: severity
          in: query
          required: false
          schema: { $ref: '#/components/schemas/NotificationSeverity' }
        - name: q
          in: query
          required: false
          schema: { type: string, maxLength: 200 }
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
        - name: from
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: false
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Filtered notification page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminNotificationPage' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      operationId: createAdminNotification
      tags: [Admin]
      summary: Create a persistent administrator notification
      security:
        - accessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAdminNotification' }
      responses:
        '201':
          description: Notification created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminNotification' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/admin/notifications/mark-read:
    post:
      operationId: markAdminNotificationsRead
      tags: [Admin]
      summary: Mark persistent notifications as read
      security:
        - accessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [ids]
              properties:
                ids:
                  type: array
                  minItems: 1
                  maxItems: 200
                  uniqueItems: true
                  items: { type: string, format: uuid }
      responses:
        '200':
          description: Number of unread notifications changed.
          content:
            application/json:
              schema:
                type: object
                required: [ok, updated]
                properties:
                  ok: { type: boolean, enum: [true] }
                  updated: { type: integer, minimum: 0 }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/admin/notifications/stream:
    get:
      operationId: streamAdminNotifications
      tags: [Admin]
      summary: Stream notification snapshots and new events
      security:
        - accessToken: []
      responses:
        '200':
          description: Server-sent event stream authenticated with the access-token header.
          content:
            text/event-stream:
              schema: { type: string }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer
      bearerFormat: Core access token
      description: Fifteen-minute Core access token returned by OTP verification or refresh.
    refreshCookie:
      type: apiKey
      in: cookie
      name: mwd_refresh
      description: Rotating HttpOnly browser refresh cookie scoped to `/api/v1/auth`.
    storageGrant:
      type: http
      scheme: bearer
      bearerFormat: Storage grant
      description: Short-lived Core-issued grant bound to an object and transfer purpose.
  parameters:
...
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Authentication or grant validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Administrator access is required.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource is absent or unavailable to the caller.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ServiceUnavailable:
      description: A required service or dependency is unavailable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    AdminUser:
      type: object
      required: [id, name, email, role, status, createdAt, quota]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, nullable: true }
        email: { type: string, format: email }
        role: { type: string, enum: [user, admin] }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        quota:
          type: object
          allOf:
            - $ref: '#/components/schemas/QuotaBalance'
          nullable: true
    AdminUserPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminUser' }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
    AdminRoleUpdate:
      type: object
      required: [id, role]
      properties:
        id: { type: string, format: uuid }
        role: { type: string, enum: [user, admin] }
    NotificationSeverity:
      type: string
      enum: [critical, warning, info, success]
    AdminNotification:
      type: object
      required: [id, title, severity, unread, createdAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 2000, nullable: true }
        severity: { $ref: '#/components/schemas/NotificationSeverity' }
        service: { type: string, maxLength: 100, nullable: true }
        unread: { type: boolean }
        createdAt: { type: string, format: date-time }
        meta:
          type: object
          additionalProperties: true
          nullable: true
    CreateAdminNotification:
      type: object
      additionalProperties: false
      required: [title, severity]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, minLength: 1, maxLength: 2000 }
        severity: { $ref: '#/components/schemas/NotificationSeverity' }
        service: { type: string, minLength: 1, maxLength: 100 }
        meta:
          type: object
          additionalProperties: true
    AdminNotificationPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminNotification' }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 200 }
        total: { type: integer, minimum: 0 }
...
    NonnegativeByteString:
      type: string
      pattern: '^(0|[1-9]\d*)$'
      description: Nonnegative byte count encoded as a decimal string.
...
    QuotaBalance:
      type: object
      required: [limitBytes, reservedBytes, committedBytes, availableBytes]
      properties:
        limitBytes: { $ref: '#/components/schemas/NonnegativeByteString' }
        reservedBytes: { $ref: '#/components/schemas/NonnegativeByteString' }
        committedBytes: { $ref: '#/components/schemas/NonnegativeByteString' }
        availableBytes: { $ref: '#/components/schemas/NonnegativeByteString' }
...